민지 12/23 최종 commit
