--- showProject/src/main/java/com/kh/show/TestClass.java
@@ -1,5 +1,0 @@
-package com.kh.show;
-
-public class TestClass {
-	//test
-}
