+++ showProject/src/main/java/com/kh/show/reservation/model/vo/Reservation.java
@@ -1,8 +1,8 @@
 package com.kh.show.reservation.model.vo;
-
 import java.sql.Date;
+import java.util.List;
 
-import com.kh.show.payments.model.vo.Payments;
+import com.kh.show.showInfo.model.vo.Show;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,4 +22,19 @@ public class Reservation {
 	private String status;//	STATUS
 	private Date createDate;//	CREATEDATE
 	private String precaution;//	PRECAUTION
+	
+	// mapper에서 collection 사용하기 위해 추가
+	private List<Show> showList;
 }
+
+
+
+
+
+
+
+
+
+
+
+
