+++ showProject/src/main/resources/mappers/showInfo-mapper.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+"https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="showInfoMapper">
+
+	<resultMap type="Show" id="showResultSet">
+		<result column="SHOW_NO" property="showNo" />
+		<result column="SHOW_NAME" property="showName" />
+		<result column="SHOW_EXPLAIN" property="showExplain" />
+		<result column="PRICE" property="price" />
+		<result column="NOTICE" property="notice" />
+		<result column="PRECAUTION" property="precaution" />
+		<result column="STATUS" property="status" />
+		<result column="HALL_NO" property="hallNo" />
+		<result column="GENRE_NO" property="genreNo" />
+		<result column="REGION_NO" property="regionNo" />
+		<result column="SHOW_START" property="showStart" />
+		<result column="SHOW_END" property="showEnd" />
+		<result column="COUNT" property="count" />
+		
+	</resultMap>	
+
+ <!-- 뮤지컬 랭킹(1~5위) count 순으로 리스트 -->
+  <select id="musicalShowList" resultMap="showResultSet">
+  	SELECT *
+  	FROM (
+  		  SELECT SHOW_NAME
+  	       		 ,SHOW_START
+  	       		 ,SHOW_END
+  	      FROM SHOW
+  	      WHERE GENRE_NO = '2'
+  	      ORDER BY COUNT DESC
+  	      )
+
+  	WHERE ROWNUM BETWEEN 1 AND 5
+  </select>
+  	       
+  
+
+	
+	
+		
+</mapper>
