+++ showProject/src/main/webapp/WEB-INF/views/main.jsp
@@ -33,8 +33,8 @@
 	<%@include file="/WEB-INF/views/common/posterslider.jsp" %> 
     <%@include file="/WEB-INF/views/common/ticketOpen.jsp" %>      
     <%@include file="/WEB-INF/views/common/ranking.jsp" %>  
-     <br><br>
-    <%@include file="/WEB-INF/views/common/today.jsp" %> 
+    <br><br>
+
